requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email-validator
python-dotenv
bcrypt
